first pass at updating w/ response to comments from Stu
